Make the order of bit packing a selectable option.
Adds an Option type that controls how a Reader or Writer behaves.  The only
option currently available is to choose which order bits should be packed into
bytes, which is either MSBFirst (the default) or LSBFirst.

All the tests ave been updated to check both possibilities, and the comments
that referred to a fixed encoding have been removed.